apply lifting state
